Refactor: split post style into a seperate file
